backend-event-listener/requirements.txt
web3==7.12.0
pymongo==4.6.1
python-dotenv==1.0.1
motor==3.3.0
redis==5.0.1